PaymentService/Core/Application/Application/PaymentProcessorFactory.cs
using Application.MercadoPago;
using Application.PaymentApplication.Dtos;
using Application.PaymentApplication.Ports;

namespace Payment.Application
{
    public class PaymentProcessorFactory : IPaymentProcessorFactory
    {
        public IPaymentProcessor GetPaymentProcessor(SupportedPaymentProviders selectedPaymentProvider)
        {
            switch (selectedPaymentProvider)
            {
                case SupportedPaymentProviders.MercadoPago:
                    return new MercadoPagoAdapter();

                default: return new NotImplementedPaymentProvider();
            }
        }
    }
}
